refactor(classifications): consolidate Moderations into Classifications resource
Aligns API structure with OpenAPI specification where moderation and
classification endpoints are tagged as 'classifiers'.

Changes:
- Moved moderation methods to Classifications resource
- Relocated Request classes from Moderations to Classifications namespace
- Consolidated duplicate request DTOs (ClassificationRequest, ChatClassificationRequest)
- Added deprecation notice to Mistral->moderations() method
- Updated all tests and examples to use classifications() method
- Removed obsolete Moderations resource class

Breaking Changes:
- Direct imports of Moderations Request classes need namespace update
- Moderations resource deprecated (still works via alias)

Migration Path:
- Change $mistral->moderations() to $mistral->classifications()
- Update imports: Requests\Moderations\* → Requests\Classifications\*

+++ src/Resource/Chat.php
@@ -20,7 +20,7 @@ class Chat extends BaseResource
         string $model = Model::mistral7b->value,
         float $temperature = 0.7,
         int $maxTokens = 2000,
-        int $topP = 1,
+        float $topP = 1.0,
         bool $safeMode = false,
         bool $stream = false,
         ?int $randomSeed = null,
